Fixing #336 and adding weight to inventory items, also redefines some constants to be based on MAX_STRING_CHARS instead of magic numbers, also improved debuginventory command with better formatting/details

+++ codemp/game/bg_items.h
@@ -155,7 +155,7 @@ typedef struct {
 	char internalName[MAX_ITEM_NAME];
 	int itemID;
 	jkgItemType_t itemType;
-	unsigned int weight;
+	float weight;
 	unsigned int maxStack;
 	std::string itemDescription;
